test(multipooler): move postgres monitor tests into postgres_monitor_test.go
Mirror the source split from the previous commit: relocate the monitor tests to
sit alongside the code they exercise. They had accumulated in
rpc_initialization_test.go (and one in manager_test.go), which no longer matches
where the code lives.

Moved verbatim (each test with its doc comment): TestDiscoverPostgresState_*,
TestDetermineRemedialAction(+_StalePrimaryDemote), TestStaleStandbyDemoteTarget,
TestTakeRemedialAction_* (incl. ReconcileRole), TestHasCompleteBackups_*,
TestStartPostgres_*, TestMonitorPostgres_*, TestPostgresStateEqual, and
TestPrimaryConnInfoDiffersFromRecorded.

Left in place: TestIsInitialized and TestHelperMethods (init/general), and
TestUpdateTopologyAfterPromotion (its subject updateTopologyAfterPromotion
remains in manager.go). Shared test helpers (e.g. newRemedialActionTestManager)
stay where they are — same-package test files all see them.
